feat #3 (release-1.1): class Aniamal created to mutualize code in common
