+++ Dog.php
@@ -1,46 +1,14 @@
 <?php
 
-class Dog
+class Dog extends Animal
 {
-    private $name = 'indéfinie';
-    private $race = 'indéfinie';
-    private $age = 0; // 0 egal age not uninformed or age < 1 like baby
     private $weight = 15;
     private const SPEEDING = 60;
 
     public function __construct()
     {
-        echo 'Vous avez adoptez un nouveau ' . strtolower(get_class()) . '.<br>';
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(string $name)
-    {
-        $this->name = $name;
-    }
-
-    public function getRace(): ?string
-    {
-        return $this->race;
-    }
-
-    public function setRace(string $race)
-    {
-        $this->race = $race;
-    }
-
-    public function getAge(): ?int
-    {
-        return $this->age;
-    }
-
-    public function setAge(int $age)
-    {
-        $this->age = $age;
+        parent::__construct();
+        echo 'Il s\'agit d\'un ' . strtolower(get_class()) . '.<br>';
     }
 
     public function getWeight(): ?int
