+++ index.php
@@ -1,6 +1,7 @@
 <?php
 declare(strict_types=1);
 
+require_once 'Animal.php';
 require_once 'Cat.php';
 require_once 'Dog.php';
